# Conflicts:
#	cli/pom.xml
#	cli/src/main/java/com/devonfw/tools/ide/common/SystemPath.java
#	cli/src/main/java/com/devonfw/tools/ide/tool/ToolCommandlet.java
#	cli/src/test/java/com/devonfw/tools/ide/context/AbstractIdeContextTest.java
#	documentation/LICENSE.adoc

+++ cli/src/main/java/com/devonfw/tools/ide/tool/GlobalToolCommandlet.java
@@ -1,5 +1,11 @@
 package com.devonfw.tools.ide.tool;
 
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+
 import com.devonfw.tools.ide.common.Tag;
 import com.devonfw.tools.ide.context.IdeContext;
 import com.devonfw.tools.ide.io.FileAccess;
@@ -8,10 +14,6 @@ import com.devonfw.tools.ide.process.ProcessContext;
 import com.devonfw.tools.ide.process.ProcessErrorHandling;
 import com.devonfw.tools.ide.repo.ToolRepository;
 import com.devonfw.tools.ide.version.VersionIdentifier;
-
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.Set;
 
 /**
  * {@link ToolCommandlet} that is installed globally.
@@ -29,6 +31,63 @@ public abstract class GlobalToolCommandlet extends ToolCommandlet {
   public GlobalToolCommandlet(IdeContext context, String tool, Set<Tag> tags) {
 
     super(context, tool, tags);
+  }
+
+  /**
+   * Performs the installation of the {@link #getName() tool} via a package manager.
+   *
+   * @param silent - {@code true} if called recursively to suppress verbose logging, {@code false} otherwise.
+   * @param commands - A {@link Map} containing the commands used to perform the installation for each package manager.
+   * @return {@code true} if the tool was newly installed, {@code false} if the tool was already installed before and
+   *         nothing has changed.
+   */
+  protected boolean installWithPackageManger(Map<PackageManager, List<String>> commands, boolean silent) {
+
+    Path binaryPath = this.context.getPath().findBinary(Path.of(getBinaryName()));
+
+    if (binaryPath != null && Files.exists(binaryPath) && !this.context.isForceMode()) {
+      IdeLogLevel level = silent ? IdeLogLevel.DEBUG : IdeLogLevel.INFO;
+      this.context.level(level).log("{} is already installed at {}", this.tool, binaryPath);
+      return false;
+    }
+
+    Path bashPath = this.context.getPath().findBinary(Path.of("bash"));
+    if (bashPath == null || !Files.exists(bashPath)) {
+      context.warning("Bash was not found on this machine. Not Proceeding with installation of tool " + this.tool);
+      return false;
+    }
+
+    PackageManager foundPackageManager = null;
+    for (PackageManager pm : commands.keySet()) {
+      if (Files.exists(this.context.getPath().findBinary(Path.of(pm.toString().toLowerCase())))) {
+        foundPackageManager = pm;
+        break;
+      }
+    }
+
+    int finalExitCode = 0;
+    if (foundPackageManager == null) {
+      context.warning("No supported Package Manager found for installation");
+      return false;
+    } else {
+      List<String> commandList = commands.get(foundPackageManager);
+      if (commandList != null) {
+        for (String command : commandList) {
+          ProcessContext pc = this.context.newProcess().errorHandling(ProcessErrorHandling.WARNING).executable(bashPath)
+              .addArgs("-c", command);
+          finalExitCode = pc.run();
+        }
+      }
+    }
+
+    if (finalExitCode == 0) {
+      this.context.success("Successfully installed {}", this.tool);
+    } else {
+      this.context.warning("{} was not successfully installed", this.tool);
+      return false;
+    }
+    postInstall();
+    return true;
   }
 
   @Override
@@ -58,17 +117,20 @@ public abstract class GlobalToolCommandlet extends ToolCommandlet {
     // download and install the global tool
     FileAccess fileAccess = this.context.getFileAccess();
     Path target = toolRepository.download(this.tool, edition, resolvedVersion);
-    Path tmpDir = fileAccess.createTempDir(getName());
-    Path downloadBinaryPath = tmpDir.resolve(target.getFileName());
-    extract(target, downloadBinaryPath);
-    if (isExtract()) {
+    Path executable = target;
+    Path tmpDir = null;
+    boolean extract = isExtract();
+    if (extract) {
+      tmpDir = fileAccess.createTempDir(getName());
+      Path downloadBinaryPath = tmpDir.resolve(target.getFileName());
+      fileAccess.extract(target, downloadBinaryPath);
       downloadBinaryPath = fileAccess.findFirst(downloadBinaryPath, Files::isExecutable, false);
     }
-    ProcessContext pc = this.context.newProcess().errorHandling(ProcessErrorHandling.WARNING)
-        .executable(downloadBinaryPath);
+    ProcessContext pc = this.context.newProcess().errorHandling(ProcessErrorHandling.WARNING).executable(executable);
     int exitCode = pc.run();
-    fileAccess.delete(tmpDir);
-    fileAccess.delete(target);
+    if (tmpDir != null) {
+      fileAccess.delete(tmpDir);
+    }
     if (exitCode == 0) {
       this.context.success("Successfully installed {} in version {}", this.tool, resolvedVersion);
     } else {
@@ -78,5 +140,4 @@ public abstract class GlobalToolCommandlet extends ToolCommandlet {
     postInstall();
     return true;
   }
-
 }
